more efficient debug

+++ upload.sh
@@ -4,9 +4,14 @@ source ~/.bash_profile
 cd ..
 
 rm -rf pigfarm.tar.gz
-tar cvf pigfarm.tar.gz --exclude='*.git' pigfarm
 
-ssh-pigfarm 'rm -rf pigfarm'
+if [ "$1" == "--essential" ]; then
+	tar cvf pigfarm.tar.gz --exclude='*.git' --exclude='node_module' --exclude='prettyphoto' --exclude='images' --exclude='fonts' pigfarm
+else
+	tar cvf pigfarm.tar.gz --exclude='*.git' pigfarm
+	ssh-pigfarm 'rm -rf pigfarm'
+fi
+
 scp -i ~/keys/amazon.jiwon.pem pigfarm.tar.gz ubuntu@ec2-52-25-245-158.us-west-2.compute.amazonaws.com:/home/ubuntu
 
 ssh-pigfarm 'tar xvf pigfarm.tar.gz'
